changes: 1. fixed (renamed) roommate tabs packge. 2. changed the roommateSearcherUser onBoarding to editProfile.
