+++ app/src/main/AndroidManifest.xml
@@ -32,9 +32,6 @@
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.Dialog.Alert" />
         <activity
-            android:name=".RoommateSearcherSetProfileActivity"
-            android:screenOrientation="portrait" />
-        <activity
             android:name=".EditProfileAlertDialog"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.Dialog.Alert" />
